[Windows] Fix Pod cannot access k8s API server service issue
When a Pod access cluster service and the selected endpoint uses
node IP(hostnetwork mode). The request packets need to be SNATed
after have been DNATed. On Windows node, antrea both applied both
DNAT and SNAT in the same ct_zone. That's not supported by OVS.

In this patch, we introduce a new ct_zone to track this kind of
SNATed connection in a different ct_zone.

Fixes: #1759

Signed-off-by: Rui Cao <rcao@vmware.com>